Supporting ncurses style
