+++ include/my_ncurses.h
@@ -72,8 +72,10 @@ my_addstr(window, str))
 #define FAST_BLINK 6
 #define REVERSE 6
 
+#define DISABLE_ATTR 20
+
 #define my_attron(attr, value) (printf("\x1B[%dm", attr + value))
-#define my_attroff(attr, value) (printf("\x1B[%dm", attr + 20))
+#define my_attroff(attr, value) (printf("\x1B[%dm", attr + DISABLE_ATTR))
 #define my_attrreset() (printf("\x1B[0m"))
